upload by meng

+++ G3_1/head.h
@@ -8,6 +8,7 @@
 #include <sqlite3.h>
 #include <signal.h>
 #include <time.h>
+#include <pthread.h>
 #include <sys/socket.h>
 #include <netinet/in.h>
 #include <arpa/inet.h>
@@ -18,14 +19,14 @@
 typedef struct sockaddr SA;
 //各种命令
 enum CMD{
-    LOGIN,  //登录
-    STAFF_CHANGE_PASSWD,//修改密码
-    STAFF_QUERY,        //员工查询
-    ADMIN_ADD_STAFF,    //管理员增加员工
-    ADMIN_DEL_STAFF,    //管理员删除员工
-    ADMIN_QUERY,        //管理员查询
-    QUIT ,            //退出
-    FAILED              //失败
+	LOGIN,  //登录
+	STAFF_CHANGE_PASSWD,//修改密码
+	STAFF_QUERY,        //员工查询
+	ADMIN_ADD_STAFF,    //管理员增加员工
+	ADMIN_DEL_STAFF,    //管理员删除员工
+	ADMIN_QUERY,        //管理员查询
+	QUIT ,            //退出
+	FAILED              //失败
 };
 
 enum USER_LEVEL{
@@ -35,21 +36,21 @@ enum USER_LEVEL{
 //员工信息结构体
 typedef struct
 {
-    int id;  //员工编号
-    char name[N];//姓名
-    char addr[N];//地址
-    int  age;//年龄
-    float salary; //工资
+	int id;  //员工编号
+	char name[N];//姓名
+	char addr[N];//地址
+	int  age;//年龄
+	int salary; //工资
 }INFO;
 //所有信息的结构体
 typedef struct{
 	int USER_LEVEL;
-    int ret;  //操作结果
-    int type;//操作类型
-    INFO info;//员工信息结构体
-    char name[N];//发送消息的户
-    char passwd[N];//用户密码在第一次登录时使用
-    char data[N];//操作成功或者失败的消息
+	int ret;  //操作结果
+	int type;//操作类型
+	INFO info;//员工信息结构体
+	char name[N];//发送消息的户
+	char passwd[N];//用户密码在第一次登录时使用
+	char data[N];//操作成功或者失败的消息
 }MSG;
 
 #endif
